fix: JsonResponse를 사용해야 하는 MiddlewareMixin 대신 BaseAuthentication 중 커스텀 형식으로 변경(DRF), CORS 헤더에 인증헤더 추가
