+++ backend/backend/settings.py
@@ -71,10 +71,6 @@ MIDDLEWARE = [
     "django.middleware.common.CommonMiddleware",
     "django.middleware.csrf.CsrfViewMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
-
-    # Add jwt-authentication-middleware
-    "login.middleware.JWTAuthenticationMiddleware",
-
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
     "whitenoise.middleware.WhiteNoiseMiddleware",
@@ -82,12 +78,11 @@ MIDDLEWARE = [
 
 # DRF 설정 추가
 REST_FRAMEWORK = {
-    'DEFAULT_AUTHENTICATION_CLASSES': (
-        'rest_framework.authentication.SessionAuthentication',
-        'rest_framework.authentication.BasicAuthentication',
+    "DEFAULT_AUTHENTICATION_CLASSES": (
+        "login.authentication.JWTAuthentication",
     ),
-    'DEFAULT_PERMISSION_CLASSES': (
-        'rest_framework.permissions.IsAuthenticated',
+    "DEFAULT_PERMISSION_CLASSES": (
+        "rest_framework.permissions.IsAuthenticated",
     ),
 }
 
@@ -99,6 +94,7 @@ CSRF_TRUSTED_ORIGINS = ["localhost:8000"]
 CORS_ALLOW_HEADERS = [
     "x-csrftoken",
     "content-type",
+    "Authorization",
 ]
 
 CORS_ALLOW_ALL_ORIGINS = True
